Rename and relocate all order resources.
