+++ app/Http/Resources/OrderResources/IndexCollection.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Resources;
+namespace App\Http\Resources\OrderResources;
 
 use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\ResourceCollection;
@@ -8,9 +8,9 @@ use Illuminate\Http\Resources\Json\ResourceCollection;
 /**
  * Here we can transform the data structure of the whole collection of order model objects, that
  * is sent to the order index page. 
- * So, this TmsOrderIndexCollection.php is using TmsOrderIndexResource.php to format the data.
+ * So, this IndexCollection.php is using IndexResource.php to format the data.
  */
-class TmsOrderIndexCollection extends ResourceCollection
+class IndexCollection extends ResourceCollection
 {
     /**
      * Transform the resource collection into an array.
@@ -23,7 +23,7 @@ class TmsOrderIndexCollection extends ResourceCollection
 
             /**
              * This is the actual data (collection of order model objects) that we send to the
-             * order index page. The data is formatted in TmsOrderIndexResource.php.
+             * order index page. The data is formatted in IndexResource.php.
              */
             'data' => $this->collection,
         ];
